chore: trigger re-create of WAF ACL logging config (#439)
During the apply of #438, the WAF ACL logging config was removed
and not recreated.  This will cause the logging config for Firehose
to be recreated.

The reason this happened is because there was a misconfiguration
where two logging destinations had been set for a WAF ACL and
only one is supported.