fix: update API gateway to invoke Lambda alias
The Lambda API is setup with provisioned concurrency attached
to the `latest` Lambda alias.  By using this alias with the API
gateway integration, we'll be able to take advantage of that
provisioned concurrency.

Currently the API gateway is always invoking the `$LATEST` version
of the function code, which means the provisioned concurrency
is not used.