feat: add RDS cluster activity stream
Add an async RDS cluster activity stream.  This will write the database activity
events to an S3 bucket, where they will be available for audting and alerting
for 7 days.

The async mode of the activity stream will prioritize database performance and
the stream will start during the next maintenance window.