fix: increase RDS activity stream decrypt Lambda resources (#1244)
Update the TF module version to increase timeout and memory available to
the decrypt Lambda function.  This will allow it to keep up with the heavier
load of the database.

Update the lifecycle policy on the activity log bucket to only keep query
logs for 3 days as this is the policy used by services with more stringent
data retention requirements.