Karpenter upgrade iam (#1288)
* feat: add EC2 launch template for EKS worker nodes

Add an EC2 launch template that enables IMDSv2 for the worker
nodes.  This is a security best practice but unfortunately not
part of the default launch template.

* fix: add security groups to network interface

* chore: add checkov suppression comment

* feat: add EC2 launch template for EKS worker nodes

Add an EC2 launch template that enables IMDSv2 for the worker
nodes.  This is a security best practice but unfortunately not
part of the default launch template.

* fix: add security groups to network interface

* chore: add checkov suppression comment

* fix: remove network interfaces from template

* Updating IAM for Karpenter to reflect latest requirements

* Reverting disk size thing

---------

Co-authored-by: Pat Heard <patrick.heard@cds-snc.ca>
Co-authored-by: Jimmy Royer <jimleroyer@gmail.com>