feat: add EC2 launch template for EKS worker nodes (#985)
* feat: add EC2 launch template for EKS worker nodes

Add an EC2 launch template that enables IMDSv2 for the worker
nodes.  This is a security best practice but unfortunately not
part of the default launch template.

* fix: add security groups to network interface

* chore: add checkov suppression comment

* feat: add EC2 launch template for EKS worker nodes

Add an EC2 launch template that enables IMDSv2 for the worker
nodes.  This is a security best practice but unfortunately not
part of the default launch template.

* fix: add security groups to network interface

* chore: add checkov suppression comment

* fix: remove network interfaces from template

---------

Co-authored-by: Ben Larabie <ben.larabie@cds-snc.ca>
Co-authored-by: Jimmy Royer <jimleroyer@gmail.com>