feat: RDS snapshot exports to Platform Data Lake (#1824)
Add the IAM role that will be used to perform cross-account S3 exports of the
cluster's automated snapshots to the data lake.

This IAM role will be assumed by an export process that runs in the data lake's
AWS account and run on a nightly schedule.  These RDS snapshot exports will
have no impact to the Notify database performance.