chore: disable GuardDuty Malware S3 object tagging (#2231)
Temporarily disable the GuardDuty S3 object tagging while the API
and DDAPI code changes are worked on. This will prevent the tagging
race condition from impacting the nightly performance testing.

Additionally, this PR removes the `moved` blocks as the TF state
migration is complete and was only needed in Staging.